+++ app/app.go
@@ -1,6 +1,7 @@
 package app
 
 import (
+	"encoding/json"
 	"io"
 	"net/http"
 	"os"
@@ -178,64 +179,33 @@ func getGovProposalHandlers() []govclient.ProposalHandler {
 	return govProposalHandlers
 }
 
-var (
-	ModuleBasics = module.NewBasicManager(
-		auth.AppModuleBasic{},
-		genutil.NewAppModuleBasic(genutiltypes.DefaultMessageValidator),
-		bank.AppModuleBasic{},
-		//capability.AppModuleBasic{},
-		staking.AppModuleBasic{},
-		distr.AppModuleBasic{},
-		gov.NewAppModuleBasic(getGovProposalHandlers()),
-		params.AppModuleBasic{},
-		crisis.AppModuleBasic{},
-		slashing.AppModuleBasic{},
-		//ibc.AppModuleBasic{},
-		//ibctm.AppModuleBasic{},
-		upgrade.AppModuleBasic{},
-		evidence.AppModuleBasic{},
-		//transfer.AppModuleBasic{},
-		vesting.AppModuleBasic{},
-		consensus.AppModuleBasic{},
-		evm.AppModuleBasic{},
-		feemarket.AppModuleBasic{},
-		authoritymodule.AppModuleBasic{},
-		lightclientmodule.AppModuleBasic{},
-		crosschainmodule.AppModuleBasic{},
-		//ibccrosschain.AppModuleBasic{},
-		observermodule.AppModuleBasic{},
-		fungiblemodule.AppModuleBasic{},
-		emissionsmodule.AppModuleBasic{},
-		groupmodule.AppModuleBasic{},
-		authzmodule.AppModuleBasic{},
-	)
+type GenesisState map[string]json.RawMessage
 
-	// module account permissions
-	maccPerms = map[string][]string{
-		authtypes.FeeCollectorName:     nil,
-		distrtypes.ModuleName:          nil,
-		stakingtypes.BondedPoolName:    {authtypes.Burner, authtypes.Staking},
-		stakingtypes.NotBondedPoolName: {authtypes.Burner, authtypes.Staking},
-		govtypes.ModuleName:            {authtypes.Burner},
-		//ibctransfertypes.ModuleName:                     {authtypes.Minter, authtypes.Burner},
-		crosschaintypes.ModuleName: {authtypes.Minter, authtypes.Burner},
-		//ibccrosschaintypes.ModuleName:                   nil,
-		evmtypes.ModuleName:                             {authtypes.Minter, authtypes.Burner},
-		fungibletypes.ModuleName:                        {authtypes.Minter, authtypes.Burner},
-		emissionstypes.ModuleName:                       nil,
-		emissionstypes.UndistributedObserverRewardsPool: nil,
-		emissionstypes.UndistributedTSSRewardsPool:      nil,
-	}
+// module account permissions
+var maccPerms = map[string][]string{
+	authtypes.FeeCollectorName:     nil,
+	distrtypes.ModuleName:          nil,
+	stakingtypes.BondedPoolName:    {authtypes.Burner, authtypes.Staking},
+	stakingtypes.NotBondedPoolName: {authtypes.Burner, authtypes.Staking},
+	govtypes.ModuleName:            {authtypes.Burner},
+	//ibctransfertypes.ModuleName:                     {authtypes.Minter, authtypes.Burner},
+	crosschaintypes.ModuleName: {authtypes.Minter, authtypes.Burner},
+	//ibccrosschaintypes.ModuleName:                   nil,
+	evmtypes.ModuleName:                             {authtypes.Minter, authtypes.Burner},
+	fungibletypes.ModuleName:                        {authtypes.Minter, authtypes.Burner},
+	emissionstypes.ModuleName:                       nil,
+	emissionstypes.UndistributedObserverRewardsPool: nil,
+	emissionstypes.UndistributedTSSRewardsPool:      nil,
+}
 
-	// module accounts that are NOT allowed to receive tokens
-	blockedReceivingModAcc = map[string]bool{
-		distrtypes.ModuleName:          true,
-		authtypes.FeeCollectorName:     true,
-		stakingtypes.BondedPoolName:    true,
-		stakingtypes.NotBondedPoolName: true,
-		govtypes.ModuleName:            true,
-	}
-)
+// module accounts that are NOT allowed to receive tokens
+var blockedReceivingModAcc = map[string]bool{
+	distrtypes.ModuleName:          true,
+	authtypes.FeeCollectorName:     true,
+	stakingtypes.BondedPoolName:    true,
+	stakingtypes.NotBondedPoolName: true,
+	govtypes.ModuleName:            true,
+}
 
 var (
 	_ runtime.AppI            = (*App)(nil)
@@ -260,6 +230,7 @@ type App struct {
 
 	mm           *module.Manager
 	sm           *module.SimulationManager
+	mb           module.BasicManager
 	configurator module.Configurator
 
 	// sdk keepers
@@ -637,6 +608,7 @@ func New(
 	govRouter.AddRoute(govtypes.RouterKey, govv1beta1.ProposalHandler).
 		AddRoute(paramproposal.RouterKey, params.NewParamChangeProposalHandler(app.ParamsKeeper)).
 		AddRoute(upgradetypes.RouterKey, upgrade.NewSoftwareUpgradeProposalHandler(app.UpgradeKeeper))
+
 	govConfig := govtypes.DefaultConfig()
 	govKeeper := govkeeper.NewKeeper(
 		appCodec,
@@ -648,11 +620,14 @@ func New(
 		govConfig,
 		authAddr,
 	)
+
 	app.GovKeeper = *govKeeper.SetHooks(
 		govtypes.NewMultiGovHooks(
 		// register governance hooks
 		),
 	)
+	// Set legacy router for backwards compatibility with gov v1beta1
+	// app.GovKeeper.SetLegacyRouter(govRouter)
 
 	// Create evidence Keeper for to register the IBC light client misbehaviour evidence route
 	evidenceKeeper := evidencekeeper.NewKeeper(
@@ -738,6 +713,8 @@ func New(
 		authzmodule.NewAppModule(appCodec, app.AuthzKeeper, app.AccountKeeper, app.BankKeeper, app.interfaceRegistry),
 	)
 
+	app.mb = ModuleBasics
+
 	// During begin block slashing happens after distr.BeginBlocker so that
 	// there is nothing left over in the validator fee pool, so as to keep the
 	// CanWithdrawInvariant invariant.
@@ -812,6 +789,10 @@ func New(
 	app.mm.RegisterInvariants(&app.CrisisKeeper)
 	app.configurator = module.NewConfigurator(app.appCodec, app.MsgServiceRouter(), app.GRPCQueryRouter())
 	app.mm.RegisterServices(app.configurator)
+
+	app.sm = module.NewSimulationManager(simulationModules(app, appCodec, skipGenesisInvariants)...)
+
+	app.sm.RegisterStoreDecoders()
 
 	// initialize stores
 	app.MountKVStores(keys)
@@ -920,7 +901,7 @@ func (app *App) ModuleAccountAddrs() map[string]bool {
 	return modAccAddrs
 }
 
-// LegacyAmino returns SimApp's amino codec.
+// LegacyAmino returns app's amino codec.
 //
 // NOTE: This is solely to be used for testing purposes as it may be desirable
 // for modules to register their own custom testing types.
@@ -983,7 +964,7 @@ func (app *App) RegisterAPIRoutes(apiSvr *api.Server, apiConfig config.APIConfig
 	nodeservice.RegisterGRPCGatewayRoutes(clientCtx, apiSvr.GRPCGatewayRouter)
 
 	// Register legacy and grpc-gateway routes for all modules.
-	ModuleBasics.RegisterGRPCGatewayRoutes(clientCtx, apiSvr.GRPCGatewayRouter)
+	app.mb.RegisterGRPCGatewayRoutes(clientCtx, apiSvr.GRPCGatewayRouter)
 
 	// register app's OpenAPI routes.
 	if apiConfig.Swagger {
@@ -1073,6 +1054,10 @@ func VerifyAddressFormat(bz []byte) error {
 // SimulationManager implements the SimulationApp interface
 func (app *App) SimulationManager() *module.SimulationManager {
 	return app.sm
+}
+
+func (app *App) BasicManager() module.BasicManager {
+	return app.mb
 }
 
 func (app *App) BlockedAddrs() map[string]bool {