+++ rpc/backend/blocks_test.go
@@ -1191,6 +1191,7 @@ func (suite *BackendTestSuite) TestHeaderByNumber() {
 	var expResultBlock *tmrpctypes.ResultBlock
 
 	_, bz := suite.buildEthereumTx()
+	validator := sdk.AccAddress(tests.GenerateAddress().Bytes())
 
 	testCases := []struct {
 		name         string
@@ -1218,6 +1219,7 @@ func (suite *BackendTestSuite) TestHeaderByNumber() {
 				height := blockNum.Int64()
 				client := suite.backend.clientCtx.Client.(*mocks.Client)
 				RegisterBlockNotFound(client, height)
+
 			},
 			false,
 		},
@@ -1245,6 +1247,7 @@ func (suite *BackendTestSuite) TestHeaderByNumber() {
 
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFeeError(queryClient)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			true,
 		},
@@ -1260,6 +1263,7 @@ func (suite *BackendTestSuite) TestHeaderByNumber() {
 
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			true,
 		},
@@ -1275,6 +1279,7 @@ func (suite *BackendTestSuite) TestHeaderByNumber() {
 
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			true,
 		},
@@ -1287,7 +1292,12 @@ func (suite *BackendTestSuite) TestHeaderByNumber() {
 			header, err := suite.backend.HeaderByNumber(tc.blockNumber)
 
 			if tc.expPass {
-				expHeader := ethrpc.EthHeaderFromTendermint(expResultBlock.Block.Header, ethtypes.Bloom{}, tc.baseFee)
+				expHeader := ethrpc.EthHeaderFromTendermint(
+					expResultBlock.Block.Header,
+					ethtypes.Bloom{},
+					tc.baseFee,
+					validator,
+				)
 				suite.Require().NoError(err)
 				suite.Require().Equal(expHeader, header)
 			} else {
@@ -1303,6 +1313,7 @@ func (suite *BackendTestSuite) TestHeaderByHash() {
 	_, bz := suite.buildEthereumTx()
 	block := tmtypes.MakeBlock(1, []tmtypes.Tx{bz}, nil, nil)
 	emptyBlock := tmtypes.MakeBlock(1, []tmtypes.Tx{}, nil, nil)
+	validator := sdk.AccAddress(tests.GenerateAddress().Bytes())
 
 	testCases := []struct {
 		name         string
@@ -1355,6 +1366,7 @@ func (suite *BackendTestSuite) TestHeaderByHash() {
 
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFeeError(queryClient)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			true,
 		},
@@ -1370,6 +1382,7 @@ func (suite *BackendTestSuite) TestHeaderByHash() {
 
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			true,
 		},
@@ -1385,6 +1398,7 @@ func (suite *BackendTestSuite) TestHeaderByHash() {
 
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			true,
 		},
@@ -1397,7 +1411,12 @@ func (suite *BackendTestSuite) TestHeaderByHash() {
 			header, err := suite.backend.HeaderByHash(tc.hash)
 
 			if tc.expPass {
-				expHeader := ethrpc.EthHeaderFromTendermint(expResultBlock.Block.Header, ethtypes.Bloom{}, tc.baseFee)
+				expHeader := ethrpc.EthHeaderFromTendermint(
+					expResultBlock.Block.Header,
+					ethtypes.Bloom{},
+					tc.baseFee,
+					validator,
+				)
 				suite.Require().NoError(err)
 				suite.Require().Equal(expHeader, header)
 			} else {
@@ -1410,6 +1429,7 @@ func (suite *BackendTestSuite) TestHeaderByHash() {
 func (suite *BackendTestSuite) TestEthBlockByNumber() {
 	msgEthereumTx, bz := suite.buildEthereumTx()
 	emptyBlock := tmtypes.MakeBlock(1, []tmtypes.Tx{}, nil, nil)
+	validator := sdk.AccAddress(tests.GenerateAddress().Bytes())
 
 	testCases := []struct {
 		name         string
@@ -1453,12 +1473,14 @@ func (suite *BackendTestSuite) TestEthBlockByNumber() {
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				baseFee := sdk.NewInt(1)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			ethtypes.NewBlock(
 				ethrpc.EthHeaderFromTendermint(
 					emptyBlock.Header,
 					ethtypes.Bloom{},
 					sdk.NewInt(1).BigInt(),
+					validator,
 				),
 				[]*ethtypes.Transaction{},
 				nil,
@@ -1479,12 +1501,14 @@ func (suite *BackendTestSuite) TestEthBlockByNumber() {
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				baseFee := sdk.NewInt(1)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			ethtypes.NewBlock(
 				ethrpc.EthHeaderFromTendermint(
 					emptyBlock.Header,
 					ethtypes.Bloom{},
 					sdk.NewInt(1).BigInt(),
+					validator,
 				),
 				[]*ethtypes.Transaction{msgEthereumTx.AsTransaction()},
 				nil,
@@ -1520,6 +1544,7 @@ func (suite *BackendTestSuite) TestEthBlockByNumber() {
 func (suite *BackendTestSuite) TestEthBlockFromTendermintBlock() {
 	msgEthereumTx, bz := suite.buildEthereumTx()
 	emptyBlock := tmtypes.MakeBlock(1, []tmtypes.Tx{}, nil, nil)
+	validator := sdk.AccAddress(tests.GenerateAddress().Bytes())
 
 	testCases := []struct {
 		name         string
@@ -1543,12 +1568,14 @@ func (suite *BackendTestSuite) TestEthBlockFromTendermintBlock() {
 			func(baseFee sdkmath.Int, blockNum int64) {
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			ethtypes.NewBlock(
 				ethrpc.EthHeaderFromTendermint(
 					emptyBlock.Header,
 					ethtypes.Bloom{},
 					sdk.NewInt(1).BigInt(),
+					validator,
 				),
 				[]*ethtypes.Transaction{},
 				nil,
@@ -1578,12 +1605,14 @@ func (suite *BackendTestSuite) TestEthBlockFromTendermintBlock() {
 			func(baseFee sdkmath.Int, blockNum int64) {
 				queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			ethtypes.NewBlock(
 				ethrpc.EthHeaderFromTendermint(
 					emptyBlock.Header,
 					ethtypes.Bloom{},
 					sdk.NewInt(1).BigInt(),
+					validator,
 				),
 				[]*ethtypes.Transaction{msgEthereumTx.AsTransaction()},
 				nil,
@@ -1657,6 +1686,8 @@ func (suite *BackendTestSuite) TestEthAndSyntheticEthBlockByNumber() {
 	msgEthereumTx, _ := suite.buildEthereumTx()
 	realTx := suite.signAndEncodeEthTx(msgEthereumTx)
 
+	validator := sdk.AccAddress(tests.GenerateAddress().Bytes())
+
 	suite.backend.indexer = nil
 	client := suite.backend.clientCtx.Client.(*mocks.Client)
 	queryClient := suite.backend.queryClient.QueryClient.(*mocks.EVMQueryClient)
@@ -1664,6 +1695,7 @@ func (suite *BackendTestSuite) TestEthAndSyntheticEthBlockByNumber() {
 	RegisterBlock(client, 1, []tmtypes.Tx{realTx, tx})
 	RegisterBlockResultsWithTxResults(client, 1, []*types.ResponseDeliverTx{{}, &txRes})
 	RegisterBaseFee(queryClient, sdk.NewInt(1))
+	RegisterValidatorAccount(queryClient, validator)
 
 	// only real should be returned
 	block, err := suite.backend.EthBlockByNumber(1)